Refactored Parser and U classes

+++ src/U.php
@@ -10,56 +10,82 @@ class U
 {
     public static function saveLog($pathToDb, $pathToLog, $timezone)
     {
-        $pdo = new \PDO('sqlite:' . $pathToDb);
+        date_default_timezone_set($timezone);
+
+        $pdo = self::getPdo($pathToDb);
 
         $lastLogTime = $pdo->prepare('SELECT updated FROM source WHERE path = :pathToLog');
         $lastLogTime->execute([':pathToLog' => $pathToLog]);
         $lastLogTime = strtotime($lastLogTime->fetchColumn());
+        if (!is_int($lastLogTime)) {
+            if ((is_numeric($lastLogTime) AND !is_float($lastLogTime)) OR is_bool($lastLogTime) OR is_null($lastLogTime))
+                $lastLogTime = intval($lastLogTime);
+            else
+                throw new \InvalidArgumentException('$timestamp parameter must be integer!');
+        }
 
         $pdo->query('UPDATE source SET last_request = :now')->execute([':now' => date('c')]);
 
         try {
-            $parser = new Parser(trim($pathToLog), $timezone);
+            $filename = trim($pathToLog);
+            if (filter_var($filename, FILTER_VALIDATE_URL) !== FALSE) {
+                $file_headers = @get_headers($filename, 1);
+                if (strpos(array_values($file_headers)[0], '200 OK') === false)
+                    throw new \RuntimeException("$filename does not exist!");
+                if (isset($file_headers['Last-Modified']))
+                    $mtime = $file_headers['Last-Modified'];
+                elseif (isset($file_headers['X-Last-Modified']))
+                    $mtime = $file_headers['X-Last-Modified'];
+                if (isset($mtime))
+                    $lastModified = strtotime($mtime);
+            } else {
+                if (!file_exists($filename))
+                    throw new \RuntimeException("$filename does not exist!");
+                $lastModified = filemtime($filename);
+            }
         } catch (\RuntimeException $e) {
             error_log($e->getMessage());
             return false;
         }
 
-        if ($parser->isUpdated($lastLogTime)) {
+        if (isset($lastModified) && $lastModified > $lastLogTime) {
             $stmt = $pdo->prepare("INSERT INTO data (logdate, level, message) VALUES (:logdate, :level, :message)");
-            foreach ($parser->parse() as $row) {
-                if ($row['logTime'] > $lastLogTime) {
-                    $stmt->execute([
-                        ':logdate' => date('c', $row['logTime']),
-                        ':level' => $row['level'],
-                        ':message' => mb_convert_encoding($row['message'], 'utf-8')
-                    ]);
+
+            $log = file($filename, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+            if (empty($log))
+                return [];
+
+            foreach ($log as $key => $string) {
+                $parser = new Parser($string);
+
+                $date = $parser->date;
+                if ($date) {
+                    if (strtotime($date) > $lastLogTime) {
+                        $message = $parser->message;
+                        if ($message) {
+                            $message = $parser->filterMessage($message);
+                            if (!empty($message)) {
+                                $stmt->execute([
+                                    ':logdate' => date('c', strtotime($date)),
+                                    ':level' => $parser->messageType,
+                                    ':message' => mb_convert_encoding($message, 'utf-8')
+                                ]);
+                            }
+                        }
+                    }
                 }
             }
-            U::deleteOldRows($pathToDb);
+
+            $pdo->query("DELETE FROM data WHERE datetime(logdate, 'localtime') < datetime('now', '-7 days', '-1 hours')");
+
+            $pathStmt = $pdo->prepare("INSERT INTO source (path) VALUES (:path)");
+            $pathStmt->execute([':path' => $pathToLog]);
+
+            $updStmt = $pdo->prepare("UPDATE source SET updated = :updated, timezone = :timezone WHERE path = :path");
+            $updStmt->execute([':path' => $pathToLog, ':updated' => date('c', $lastModified), ':timezone' => $timezone]);
         }
 
-        $pathStmt = $pdo->prepare("INSERT INTO source (path) VALUES (:path)");
-        $pathStmt->execute([':path' => $pathToLog]);
-
-        $updStmt = $pdo->prepare("UPDATE source SET updated = :updated, timezone = :timezone WHERE path = :path");
-        $updStmt->execute([':path' => $pathToLog, ':updated' => date('c', $parser->lastModified), ':timezone' => $timezone]);
         return true;
-    }
-
-    private static function deleteOldRows($pathToDb)
-    {
-        $pdo = new \PDO('sqlite:' . $pathToDb);
-        $pdo->query("DELETE FROM data WHERE datetime(logdate, 'localtime') < datetime('now', '-7 days', '-1 hours')");
-    }
-
-    public static function jsonOut($data, $options = 0)
-    {
-        header('Content-Type: application/json');
-        $json = json_encode($data, $options);
-        if ($json === false)
-            $json = ['error' => json_last_error_msg()];
-        echo $json;
     }
 
     /**
@@ -71,5 +97,14 @@ class U
         $pdo = new \PDO('sqlite:' . $pathToDb);
         new Migration($pdo);
         return $pdo;
+    }
+
+    public static function jsonOut($data, $options = 0)
+    {
+        header('Content-Type: application/json');
+        $json = json_encode($data, $options);
+        if ($json === false)
+            $json = ['error' => json_last_error_msg()];
+        echo $json;
     }
 }
